student create query run

+++ 03.SQLite-Multiple-(Three)-Tables-CRUD/app/src/main/java/com/hellohasan/sqlite_multiple_three_tables_crud/model/Student.java
@@ -3,14 +3,14 @@ package com.hellohasan.sqlite_multiple_three_tables_crud.model;
 public class Student {
     private int id;
     private String name;
-    private int registrationNumber;
+    private long registrationNumber;
     private String phone;
     private String email;
 
     public Student() {
     }
 
-    public Student(int id, String name, int registrationNumber, String phone, String email) {
+    public Student(int id, String name, long registrationNumber, String phone, String email) {
         this.id = id;
         this.name = name;
         this.registrationNumber = registrationNumber;
@@ -34,11 +34,11 @@ public class Student {
         this.name = name;
     }
 
-    public int getRegistrationNumber() {
+    public long getRegistrationNumber() {
         return registrationNumber;
     }
 
-    public void setRegistrationNumber(int registrationNumber) {
+    public void setRegistrationNumber(long registrationNumber) {
         this.registrationNumber = registrationNumber;
     }
